Remove the redundant columns from db that can be calculated

+++ src/StoreApi/Entities/OrderItem.cs
@@ -4,7 +4,9 @@
     {
         public Guid Id { get; set; }
         public int Quantity { get; set; }
-        public int Price { get; set; }
+        // Including price here adds redundancy but it's useful in case of a price
+        // change in the future
+        public decimal Price { get; set; } 
         public Guid ProductId { get; set; }
         public Product? Product { get; set; }
         public Guid OrderId { get; set; }
